add domain models for match status, tie breaks, and sets; refactor player and league identifiers

+++ src/main/java/com/github/lbrcic4219rn/dddtennisleague/application/standing/dto/StandingEntryDto.java
@@ -3,7 +3,7 @@ package com.github.lbrcic4219rn.dddtennisleague.application.standing.dto;
 public record StandingEntryDto(
         String playerId,
         int rank,
-        int points,
+        double points,
         int wins,
         int losses,
         int setsWon,
